feat(parser-variables): added parser support for variable assignment/access (#9)
Updated the parser to now be able to support building ASTs for variable assignment and variable access. Variable assignment works without needing any keywords, just like in Python. It also allows for variable assignment to be done within arithmetic expressions, for example: 1 + (variable = 1) + 2. Note the brackets, which are needed for the precedence of variable assignment to take over and not cause any syntax errors. There has also been a refactoring of the ParseResult class to fix syntax errors being overwritten as the parser creates and returns the AST as a ParseResult from the recursive calls of the grammar functions. Finally, the lexer has also been refactored to allow for variable names with numbers such as variable123 and also variables with underscores such as variable_name or _variable. Unit tests were also written for all these new additions.