feat(interpreter-evaluation): added interpreter evaluation for all current source code that can be parsed (#11)
The interpreter evaluator is now able to traverse the abstract syntax tree made by the parser and produce relevant shell and console outputs (console outputs are all empty at the moment since the language has not got support for the planned pre-built output function yet). The evaluator is now able to interpret and produce outputs for arithmetic expressions, variable assignments, variable access and comparison and logical operator expressions. There is also a RuntimeError class for handling runtime errors which are relevant in this phase of the interpreting. There is also support for a single data type known as the NumberType which is used for both numbers and the underlying representation of the TRUE and FALSE keywords with values of 1 and 0 respectively. A Context class was also built to provide access to the symbol table which currently only works with the global scope of the source code providing pi as a pre-built global value as the keyword PI as well as TRUE and FALSE. It is also planned to be used to handle local scopes for when functions are implemented in the language.