feat(functions): added support for function definition/calls (#14)
The language now supports functions definitions and functions calls. This is the following syntax for function definitions (second bullet point demonstrates anonymous functions):

- function [function-name] ([args]) begin [body] end
- function ([args]) begin [body] end

This is the syntax for function calls:

- [function-name] ([args])

The language also handles runtime errors such as mismatching number of arguments passed and incorrect syntax. There has also been some refactoring of the data types of the language to include ValueType which is the parent data type of all other data types in Pseudo. This is to prevent illegal operation errors such as adding a NumberType with a FunctionType. SymbolTable was also refactored to allow a parent SymbolTable to be passed to the constructor for local scope support.