feat(strings): added basic string data type support (#15)
The language now supports string data types and a couple of basic functionalities for strings. The syntax for a string is:

- "[string-text]"

Strings in Pseudo also support the tab character '\t' and the newline character '\n'. They also support the escape character \ to allow strings such as 'string with double quote "' to be written in pseudo as "string with double quote \"". There is also string concatenation and multiplication supported such as:

- "hello " + "world" which gives "hello world".
- "hello " * 3 which gives "hello hello hello ".
- "hello" * 0 which gives "".
- "hello" * -1 which gives a runtime error for using negative multiplier.

There is also error messages for missing double quotes in strings. Single quotes are not supported as string delimiters.