feat(multi-line-statements): added multi-line code support + block code support (#19)
The language now supports multi-line statements of code which uses the newline character as a delimiter, e.g:

- x = 1\ny = 2\noutput(toString(y))

The language also now handles empty code and code that has just newline characters by outputting an empty value to the shell/console. The language also supports block statements inside some of the supported language structures such as: if/elif/else statements, for/while loops and functions. Examples for each are shown below:

- if 10 > 0 then\nstring = "success"\noutput(string)\nelse\noutput("fail")\nend
- for i = 1 to 10 loop\nx = i * 2\noutput(toString(x))\nend
- x = 1\nfinal = 3\nwhile x < final loop\nx = x + 1\nend
- function doubleXTripleY(x, y)\nx = x * 2\ny = y * 3\noutput(toString(x))\noutput(toString(y))\nend\ndoubleXTripleY(1, 2)

There has also been a change made to the single-lined version of these statements, where the end keyword is no longer required at the end of the single-lined statement.